requirements.txt
gradio
PyMuPDF
Pillow
huggingface
torch
transformers